+++ src/borg/repository.py
@@ -21,6 +21,7 @@ from .helpers import ProgressIndicatorPercent
 from .helpers import bin_to_hex
 from .helpers import hostname_is_unique
 from .helpers import secure_erase, truncate_and_unlink
+from .helpers import Manifest
 from .locking import Lock, LockError, LockErrorT
 from .logger import create_logger
 from .lrucache import LRUCache
@@ -416,7 +417,7 @@ class Repository:
             self.lock.release()
             self.lock = None
 
-    def commit(self, save_space=False):
+    def commit(self, save_space=False, compact=True, cleanup_commits=False):
         """Commit transaction
         """
         # save_space is not used anymore, but stays for RPC/API compatibility.
@@ -426,8 +427,17 @@ class Repository:
             raise exception
         self.check_free_space()
         self.log_storage_quota()
-        self.io.write_commit()
-        if not self.append_only:
+        segment = self.io.write_commit()
+        self.segments.setdefault(segment, 0)
+        self.compact[segment] += LoggedIO.header_fmt.size
+        if compact and not self.append_only:
+            if cleanup_commits:
+                # due to bug #2850, there might be a lot of commit-only segment files.
+                # this is for a one-time cleanup of these 17byte files.
+                for segment, filename in self.io.segment_iterator():
+                    if os.path.getsize(filename) == 17:
+                        self.segments[segment] = 0
+                        self.compact[segment] = LoggedIO.header_fmt.size
             self.compact_segments()
         self.write_index()
         self.rollback()
@@ -460,7 +470,7 @@ class Repository:
                 raise
             self.prepare_txn(self.get_transaction_id())
             # don't leave an open transaction around
-            self.commit()
+            self.commit(compact=False)
             return self.open_index(self.get_transaction_id())
 
     def prepare_txn(self, transaction_id, do_cleanup=True):
@@ -676,6 +686,8 @@ class Repository:
             nonlocal unused
             # commit the new, compact, used segments
             segment = self.io.write_commit(intermediate=intermediate)
+            self.segments.setdefault(segment, 0)
+            self.compact[segment] += LoggedIO.header_fmt.size
             logger.debug('complete_xfer: wrote %scommit at segment %d', 'intermediate ' if intermediate else '', segment)
             # get rid of the old, sparse, unused segments. free space.
             for segment in unused:
@@ -951,7 +963,6 @@ class Repository:
                     if current_index.get(key, (-1, -1)) != value:
                         report_error('Index mismatch for key {}. {} != {}'.format(key, value, current_index.get(key, (-1, -1))))
         if repair:
-            self.compact_segments()
             self.write_index()
         self.rollback()
         if error_found:
@@ -1240,8 +1251,8 @@ class LoggedIO:
     def segment_filename(self, segment):
         return os.path.join(self.path, 'data', str(segment // self.segments_per_dir), str(segment))
 
-    def get_write_fd(self, no_new=False, raise_full=False):
-        if not no_new and self.offset and self.offset > self.limit:
+    def get_write_fd(self, no_new=False, want_new=False, raise_full=False):
+        if not no_new and (want_new or self.offset and self.offset > self.limit):
             if raise_full:
                 raise self.SegmentFull
             self.close_segment()
@@ -1453,7 +1464,7 @@ class LoggedIO:
         if data_size > MAX_DATA_SIZE:
             # this would push the segment entry size beyond MAX_OBJECT_SIZE.
             raise IntegrityError('More than allowed put data [{} > {}]'.format(data_size, MAX_DATA_SIZE))
-        fd = self.get_write_fd(raise_full=raise_full)
+        fd = self.get_write_fd(want_new=(id == Manifest.MANIFEST_ID), raise_full=raise_full)
         size = data_size + self.put_header_fmt.size
         offset = self.offset
         header = self.header_no_crc_fmt.pack(size, TAG_PUT)
@@ -1463,7 +1474,7 @@ class LoggedIO:
         return self.segment, offset
 
     def write_delete(self, id, raise_full=False):
-        fd = self.get_write_fd(raise_full=raise_full)
+        fd = self.get_write_fd(want_new=(id == Manifest.MANIFEST_ID), raise_full=raise_full)
         header = self.header_no_crc_fmt.pack(self.put_header_fmt.size, TAG_DELETE)
         crc = self.crc_fmt.pack(crc32(id, crc32(header)) & 0xffffffff)
         fd.write(b''.join((crc, header, id)))
@@ -1471,14 +1482,11 @@ class LoggedIO:
         return self.segment, self.put_header_fmt.size
 
     def write_commit(self, intermediate=False):
+        # Intermediate commits go directly into the current segment - this makes checking their validity more
+        # expensive, but is faster and reduces clobber. Final commits go into a new segment.
+        fd = self.get_write_fd(want_new=not intermediate)
         if intermediate:
-            # Intermediate commits go directly into the current segment - this makes checking their validity more
-            # expensive, but is faster and reduces clobber.
-            fd = self.get_write_fd()
             fd.sync()
-        else:
-            self.close_segment()
-            fd = self.get_write_fd()
         header = self.header_no_crc_fmt.pack(self.header_fmt.size, TAG_COMMIT)
         crc = self.crc_fmt.pack(crc32(header) & 0xffffffff)
         fd.write(b''.join((crc, header)))